requirements.txt
ccxt
pandas
ta
numpy
pytz
mplfinance
matplotlib
python-telegram-bot